Improve concurrency in bgzf_mt
Instead of periodically stalling the writer to compress buffered data,
run multithreaded compression on a 'background' buffer while continuing
to ingest data into a 'foreground' buffer. When the foreground is full
and the current round of background compression is complete, write out
the results and swap the foreground & background.

Throughput with multiple cores should increase under most practical
circumstances, because the compression worker threads don't have to wait
as long for the next batch of data from the writer. However, there are
plausible circumstances in which this change will decrease throughput
(high compression level, very fast writer, few cores).

This doubles the memory usage associated with bgzf_mt, which was however
modest to begin with (~16MB*n_threads, based on the suggested upper
limit of n_sub_blks).